+++ djiffy/apps.py
@@ -3,5 +3,5 @@ from django.apps import AppConfig
 
 class DjiffyConfig(AppConfig):
     name = 'djiffy'
-    verbose_name = 'Djiffy IIIF Content'
+    verbose_name = 'IIIF Content'