+++ djiffy/management/commands/import_manifest.py
@@ -15,10 +15,12 @@ class Command(BaseCommand):
     def add_arguments(self, parser):
         parser.add_argument('path', nargs='+',
             help='One or more IIIF Collections or Manifests as file or URL')
+        parser.add_argument('--update', action='store_true',
+            help='Update previously imported manifests')
 
     def handle(self, *args, **kwargs):
         ManifestImporter(stdout=self.stdout, stderr=self.stderr,
-                         style=self.style) \
+                         style=self.style, update=kwargs['update']) \
             .import_paths(kwargs['path'])